requirements.txt
aiofiles==24.1.0
aiogram==3.20.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.0b0
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
certifi==2025.4.26
charset-normalizer==3.4.2
distro==1.9.0
frozenlist==1.6.0
h11==0.16.0
httpcore==1.0.9
httpx==1.0.0b0
idna==3.10
jiter==0.10.0
joblib==1.5.0
magic-filter==1.0.12
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.4.4
numpy==2.2.6
openai==1.53.0
packaging==25.0
propcache==0.3.1
psycopg2-binary==2.9.10
pydantic==2.11.4
pydantic_core==2.34.1
python-dotenv==1.1.0
pytz==2025.2
redis==6.1.0
rfc3986==2.0.0
scikit-learn==1.5.2
scipy==1.15.3
sniffio==1.3.1
SQLAlchemy==2.0.41
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
yarl==1.20.0
